Limit the search for the next valid object within the next 10000 bytes
When searching for the next object on the latest patch, it would find a
match for Gaia objects in a new unit that is part of the next DLC, which
happens to have a matching pattern. To avoid this issue, limit the
search within a reasonable size (10'000 bytes is way more than the size
of objects, which are normally less than 1'000 bytes long).